Fixed a stupid mistake that the order of applying DCT and IDCT is reversed
Fixed a stupid mistake that the order of applying DCT and IDCT is reversed (it should be DCT->filter->IDCT), which leads to over-filtering in fine structures and produces artifacts such as blocking, ringing and desaturation. Now it acctually filters as expected.
Adjust hard-thresholding table to fit the unnormalized DCT, to be equivalent to filtering in orthogonal transform
Default value of hard_thr changed from 2.3 to 2.8 for profile="vn", from 2.2 to 2.7 for other profiles

Some function arguments changed to const reference
Some read-only references in function arguments changed to const
Some read-only member functions changed to const

Cosmetics